test: Adding generateToken unit test
