+++ packages/server/mutation/setCookie.js
@@ -1,0 +1,8 @@
+const setCookie = ({ tokenName, token, res }) => {
+  res.cookie(tokenName, token, {
+    httpOnly: true,
+    secure: process.env.NODE_ENV === 'production',
+  })
+}
+
+module.exports = setCookie
